autobuild: Add target for autobuild
For the time being we will do auto builds from my repo based on the
"released" branch. This change adds a Dockerfile.autobuild which can be
used with dockerhubs autobuild system. It also adds a target in the
Makefile which can be used to create an image in the same way.

Signed-off-by: Steve Milner <smilner@redhat.com>